refactor: OrderInfo -> record 객체로 변경 및 코드 수정 적용
