+++ src/main/java/kr/hhplus/be/server/order/application/dto/OrderInfo.java
@@ -1,18 +1,28 @@
 package kr.hhplus.be.server.order.application.dto;
 
-import lombok.Builder;
-import lombok.Data;
+import kr.hhplus.be.server.coupon.domain.entity.Coupon;
+import kr.hhplus.be.server.exception.ErrorCode;
+import kr.hhplus.be.server.exception.InvalidRequestException;
+import kr.hhplus.be.server.order.domain.entity.Order;
 
 import java.time.LocalDateTime;
 
-@Data
-@Builder
-public class OrderInfo {
-    private Long orderId;
-    private Long userId;
-    private Long totalPrice;
-    private Long discountAmount;
-    private LocalDateTime orderDate;
-    private Long couponId;
+public record OrderInfo(
+        Long orderId,
+        Long userId,
+        Long totalPrice,
+        Long discountAmount,
+        LocalDateTime orderDate,
+        Long couponId
 
+) {
+    public static OrderInfo from(Order order, Coupon coupon) {
+        if (order == null) throw new InvalidRequestException(ErrorCode.NOT_NULL, "order");
+        return new OrderInfo(order.getId(),
+                order.getUserId(),
+                order.getTotalAmount(),
+                coupon.discountPrice(order.getTotalAmount()),
+                order.getOrderDate(),
+                order.getCouponId());
+    }
 }
